If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -14,11 +14,12 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array != NULL && action != NULL)
 	{
 		i = 0;
+
 		while (i < size)
 		{
 			action(array[i]);
